+++ business_logic/Entities/Auto.cs
@@ -11,6 +11,6 @@ namespace business_logic.Entities
         public string Model { get; set; }
         public string Color { get; set; }
         public string Number { get; set; }
-        public decimal LoadCarryingCapacity { get; set; }
+        public double LoadCarryingCapacity { get; set; }
     }
 }
